Corrige migrations e sincroniza banco de produção

+++ db/migrate/20250728230529_replace_status_string_with_integer_in_fretes.rb
@@ -1,6 +1,6 @@
 class ReplaceStatusStringWithIntegerInFretes < ActiveRecord::Migration[7.1]
   def change
-    remove_column :fretes, :status, :string
-    add_column :fretes, :status, :integer, default: 0
+    remove_column :fretes, :status, :string if column_exists?(:fretes, :status, :string)
+    add_column :fretes, :status, :integer, default: 0 unless column_exists?(:fretes, :status, :integer)
   end
 end
